composition file, and derive_global
