+++ vivarium/composites/gfp_expression.py
@@ -87,7 +87,8 @@ def generate_gfp_compartment(config):
     return compose_gene_expression(gfp_config)
 
 if __name__ == '__main__':
-    from vivarium.actor.process import load_compartment, simulate_compartment, convert_to_timeseries
+    from vivarium.actor.process import load_compartment, simulate_compartment
+    from vivarium.actor.composition import convert_to_timeseries
 
     out_dir = os.path.join('out', 'tests', 'gfp_expression_composite')
     if not os.path.exists(out_dir):
