+++ vivarium/processes/convenience_kinetics.py
@@ -4,8 +4,8 @@ import os
 
 from scipy import constants
 
-from vivarium.actor.process import Process, convert_to_timeseries, plot_simulation_output, \
-    simulate_process_with_environment
+from vivarium.actor.process import Process
+from vivarium.actor.composition import simulate_process_with_environment, convert_to_timeseries, plot_simulation_output
 from vivarium.utils.kinetic_rate_laws import KineticFluxModel
 from vivarium.utils.dict_utils import tuplify_role_dicts
 from vivarium.utils.units import units
